Finalized parameters for approximation

+++ src/p-approximations.cc
@@ -15,7 +15,7 @@ std::vector<double> _get_G(
         double K_m,
         std::vector<double> P_k,
         std::vector<double> S_k_half,
-        std::vector<int> alpha,
+        std::vector<double> alpha,
         std::vector<double> g
 ) {
     std::vector<double> G;
@@ -71,7 +71,7 @@ std::vector<double> getApproximatePkHalf(
         std::vector<double> S_k_half,
         std::vector<double> D_p,
         std::vector<double> x,
-        std::vector<int> alpha,
+        std::vector<double> alpha,
         std::vector<double> g,
         double t_step,
         double V_max,
