+++ src/s-approximations.h
@@ -4,7 +4,7 @@ std::vector<double> getApproximateSkHalf(
         std::vector<double> S_k,
         std::vector<double> D_s,
         std::vector<double> x,
-        std::vector<int> alpha,
+        std::vector<double> alpha,
         std::vector<double> f,
         double t_step,
         double V_max,
